pyproject.toml
[tool.poetry]
name = "st-pages"
version = "0.3.5"
description = "An experimental version of Streamlit Multi-Page Apps"
authors = ["Zachary Blackwood <zachary@streamlit.io>"]
readme = "README.md"
packages = [{include = "st_pages", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.9.7 || >3.9.7,<4.0"
streamlit = ">=1.10.0"
numpy = "^1.23.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
black = "^22.8.0"
isort = "^5.10.1"
mypy = "^0.971"
tox = "^3.27.1"
pytest-playwright = "^0.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"
line_length = 88
skip = ["./.venv", "./direnv", ".env"]

[tool.black]
exclude = '''
(
  /(
      \.vscode
    | \.git
    | \.pytest_cache
    | \.mypy_cache
    | \.venv
    | \.env
    | \.direnv
  )/
)
'''
include = '\.pyi?$'
line-length = 88

[tool.mypy]
files = [
  "**/*.py",
]
follow_imports = "silent"
ignore_missing_imports = true
scripts_are_modules = true
python_version = 3.9
